+++ plugins/org.projectforge.plugins.datatransfer/src/main/kotlin/org/projectforge/plugins/datatransfer/rest/DataTransferPersonalBoxPageRest.kt
@@ -56,7 +56,7 @@ class DataTransferPersonalBoxPageRest : AbstractDynamicPageRest() {
 
   @GetMapping("dynamic")
   fun getForm(request: HttpServletRequest, @RequestParam("userId") userIdString: String?): FormLayoutData {
-    val userId = NumberHelper.parseInteger(userIdString) ?: getUserPref().userId
+    val userId = NumberHelper.parseLong(userIdString) ?: getUserPref().userId
     val dto = DataTransferPersonalBox()
     userId?.let {
       dto.user = User.getUser(userId)
@@ -80,7 +80,7 @@ class DataTransferPersonalBoxPageRest : AbstractDynamicPageRest() {
     )
   }
 
-  private fun getLayout(id: Int?, userId: Int?): UILayout {
+  private fun getLayout(id: Long?, userId: Long?): UILayout {
     val layout = UILayout("plugins.datatransfer.title.heading")
     layout.add(
       UIFieldset(title = "plugins.datatransfer.personalBox")
@@ -118,7 +118,7 @@ class DataTransferPersonalBoxPageRest : AbstractDynamicPageRest() {
     return layout
   }
 
-  private fun ensurePersonalBox(dto: DataTransferPersonalBox): Int? {
+  private fun ensurePersonalBox(dto: DataTransferPersonalBox): Long? {
     val userId = dto.user?.id ?: return null
     val pfUser = userGroupCache.getUser(userId) ?: return null
     val dbo = dataTransferAreaDao.ensurePersonalBox(userId)
@@ -129,7 +129,7 @@ class DataTransferPersonalBoxPageRest : AbstractDynamicPageRest() {
     return dbo.id
   }
 
-  class PersonalBoxUserPref(var userId: Int? = null)
+  class PersonalBoxUserPref(var userId: Long? = null)
 
   private fun getUserPref(): PersonalBoxUserPref {
     return userPrefService.ensureEntry("datatransfer", "personalbox", PersonalBoxUserPref())
